+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 ### Features
 
 * Allow to compile castor with PHP 8.5
+* Castor static binaries are now compiled with PHP 8.5
 
 ### Minor
 
-* Rework how repacking is done. It's no longer needed to require jolicode/castor before repacking.
+* Rework how repacking is done. It's no longer needed to require `jolicode/castor` before repacking
+* Update SPC version to v2.8.2
 
 ## 1.2.0 (2026-02-10)